backend/requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asn1crypto==1.5.1
asyncpg==0.31.0
boto3==1.42.44
botocore==1.42.44
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.4
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
filelock==3.20.3
fonttools==4.62.1
fsspec==2026.2.0
google-genai==1.62.0
greenlet==3.3.1
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
platformdirs==4.5.1
polars==1.39.0
polars-runtime-32==1.39.0
prometheus_client==0.24.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
rsa==4.9.1
s3transfer==0.16.0
scipy==1.17.1
sentry-sdk==2.52.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==4.2.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.46
sqlmodel==0.0.32
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
tomlkit==0.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.21
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==15.0.1
